move all classifier code to a new 'classifiers' submodule
